rename enum module to fix import issues
