+++ pyckson/serializer.py
@@ -1,14 +1,14 @@
 from enum import Enum
 
-from pyckson.const import BASIC_TYPES, LIST_TYPES
+from pyckson.const import LIST_TYPES
 from pyckson.helpers import get_model, is_base_type
 
 
-def serialize_enum(cls, value):
+def serialize_enum(value):
     return value.name
 
 
-def serialize_list(sub_type, value):
+def serialize_list(value):
     return [serialize(sub_value) for sub_value in value]
 
 
@@ -22,7 +22,7 @@ def serialize_class(obj):
         elif value is None:
             raise ValueError('attribute {} of {} is None but not marked as optional'.format(attribute.python_name, obj))
         elif type(attribute.attr_type) in LIST_TYPES:
-            result[attribute.json_name] = serialize_list(attribute.attr_type.sub_type, value)
+            result[attribute.json_name] = serialize_list(value)
         else:
             try:
                 result[attribute.json_name] = serialize(value)
@@ -37,6 +37,6 @@ def serialize(obj):
     if is_base_type(obj):
         return obj
     elif issubclass(obj.__class__, Enum):
-        return serialize_enum(obj.__class__, obj)
+        return serialize_enum(obj)
     else:
         return serialize_class(obj)
